Separate the decision-related code into a new module
... so that it can be more easily reused by other components
