+++ packages/moxb/src/index.ts
@@ -1,3 +1,4 @@
 export * from './impl';
 export * from './types';
 export * from './routing';
+export * from './decision';
